+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 -   Changed gulp prettier tasks log level to warn
+-   Removed webpack-stream
+-   Updated gulp scripts tasks to use gulp-shell
+-   Updated webpack config to only output assets and errors
 
 ### Added